Crate odilia_input

Modules

proxy 🔒

Functions

get_file_paths 🔒

get_log_file_name 🔒

handle_event 🔒

or_cancel 🔒

setup_input_server

Open a socket to handle Odilia’s input events from an input server. This function will exit upon the expiry of the cancellation token passed in.

sr_event_receiver

Receives odilia_common::events::ScreenReaderEvent structs, then creates a stream of futures that need to be awaited/spawned by the caller onto the executor. Normally, the best way to do this is like so:

sr_event_receiver_inner 🔒

This contains the logic for sr_event_receiver. Since formatting doesn’t work inside macros, we compute the result here, and send it back up to be yielded/break the loop.