odilia_input

Function sr_event_receiver

pub fn sr_event_receiver(
    listener: UnixListener,
    event_sender: Sender<ScreenReaderEvent>,
    shutdown: CancellationToken,
) -> impl Stream<Item = BoxFuture<'static, ()>>

Receives odilia_common::events::ScreenReaderEvent structs, then creates a stream of futures that need to be awaited/spawned by the caller onto the executor. Normally, the best way to do this is like so:

use odilia_input::sr_event_receiver;
// use smol::spawn or tokio::spawn
let listener = UnixListener::bind("/some/path/here")
    .expect("Valid listener");
let (sender, _receiver) = bounded(128);
let ct = CancellationToken::new();
// For tokio; for async-io based executors, remember to call .detach()
let stream = sr_event_receiver(listener, sender, ct)
    .for_each(|fut| spawn(fut));

If the cancellation token is triggered, this stream will finish.