odilia_input

Function sr_event_receiver

pub async fn sr_event_receiver(
    event_sender: Sender<ScreenReaderEvent>,
    shutdown: CancellationToken,
) -> Result<()>

Receives odilia_common::events::ScreenReaderEvent structs, then sends them over the event_sender socket. This function will exit upon the expiry of the cancellation token passed in.

Errors

This function will return an error type if the same function is already running. This is checked by looking for a file on disk. If the file exists, this program is probably already running. If there is no way to get access to the directory, then this function will call exit(1); TODO: should probably return a result instead.