odilia_input

Function sr_event_receiver_inner

pub(crate) async fn sr_event_receiver_inner(
    listener: &UnixListener,
    event_sender: &Sender<ScreenReaderEvent>,
    shutdown: &CancellationToken,
) -> Result<BoxFuture<'static, ()>, ControlFlow<()>>

This contains the logic for sr_event_receiver. Since formatting doesn’t work inside macros, we compute the result here, and send it back up to be yielded/break the loop.